cloud_scripts/requirements.txt
requests
google.cloud.firestore